backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib
python-jose[cryptography]
python-multipart
email-validator
pillow
